LayoutTests/platform/chromium/accessibility/attributed-string-includes-highlighting-expected.txt

item0
item 1
item2
This verifies that using the mark tag will add the appropriate attribute to the attributed string and that we can find elements with highlighting using searching.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS item1.attributedStringForTextMarkerRangeContainsAttribute('AXHighlight', markerRange) is true
PASS item2.attributedStringForTextMarkerRangeContainsAttribute('AXHighlight', markerRange) is false
PASS resultElement.isEqual(item1.childAtIndex(0)) is true
PASS successfullyParsed is true

TEST COMPLETE
